mod: int -> ll
commit f21104cacf7f92eb98d2c4061c6c5e0293c38472
Merge: c1711c1 62f26ae
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Wed May 1 22:05:06 2024 -0600

    Merge branch 'mod_to_ll' of github.com:programming-team-code/programming_team_code into mod_to_ll

commit 62f26ae8b015ca5209352bf5c93de3754e520576
Author: GitHub <noreply@github.com>
Date:   Thu May 2 04:05:51 2024 +0000

    [auto-verifier] verify commit 80488934db64afad34a0007ec1a77248e7709ecc

commit c1711c1a58a24dfe8dc90e42275fe428e8c9916f
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Wed May 1 22:05:00 2024 -0600

    fix another test

commit 80488934db64afad34a0007ec1a77248e7709ecc
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Wed May 1 21:59:59 2024 -0600

    fix test

commit c0104ccbc8aa62eacd68f71b03d1a32684678d3c
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Wed May 1 21:58:41 2024 -0600

    fix doxygen

commit 46a05a6fb457dcfd78d9f41f6f7da58c29b124e1
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Wed May 1 21:57:49 2024 -0600

    ll mod

+++ library/convolution/gcd_convolution.hpp
@@ -1,6 +1,6 @@
 /** @file */
 #pragma once
-const int mod = 998'244'353;
+const ll mod = 998'244'353;
 /**
  * @param a,b arrays of the same length
  * @returns array `c` where `c[k]` = the sum of (a[i] * b[j]) for all pairs
