+++ library/math/binary_exponentiation_mod.hpp
@@ -1,6 +1,6 @@
 /** @file */
 #pragma once
-const int mod = 998'244'353;
+const ll mod = 998'244'353;
 /**
  * @see https://github.com/kth-competitive-programming/kactl /blob/main/content/number-theory/ModPow.h
  * @param b base
